Check*.cmake: add support for imported targets in CMAKE_REQUIRED_LIBRARIES
This commit adds the function
HANDLE_IMPORTED_TARGETS_IN_CMAKE_REQUIRED_LIBRARIES(), which
iterates over CMAKE_REQUIRED_LIBRARIES and checks each entry whether
it is the name of an imported target.
If this is the case, it queries the location of this target
and puts this in the output variable instead.
This is in use by KDE since 2 years.

Alex