-don't pull in CheckTypeSize.cmake from the cmake which is being built
We can be sure that at least cmake 2.6.3 is used when building cmcurl.
This means we always get in the first branch of the if().
I think it is not a good idea to pull a cmake module from the cmake
which is being built in, since this may use features which are not
supported in the cmake which is used to build cmake (e.g. CMAKE_CURRENT_LIST_DIR
which does not exist in cmake 2.6.3 which is the minimum for cmcurl).

A bit further below there is anyway code to handle the case that cmake is
older than 2.8.0, so it should be ok.

Alex