Support building shared libraries or modules without soname (#13155)
Add a boolean target property NO_SONAME which may be used to disable soname for
the specified shared library or module even if the platform supports it. This
property should be useful for private shared libraries or various plugins which
live in private directories and have not been designed to be found or loaded
globally.

The direct CMAKE_SHARED_LIBRARY_SONAME_${LANG}_FLAG rule variable is replaced
with a conditional SONAME_FLAG which is expanded to the value of the
CMAKE_SHARED_LIBRARY_SONAME_${LANG}_FLAG definition as long as soname supports
is enabled for the target in question.