Revert topic 'module-no-soname' (#13155)
Revert commit 6ad39679 (OS X: Fix Framework install-name after NO_SONAME
change, 2012-04-23), commit 69626f02 (Test NO_SONAME property,
2012-04-23), and commit 2d951c1d (Support building shared libraries or
modules without soname, 2012-04-22).  The topic has some subtle
problems.

(1) The implementation on the Ninja generator does not work.  The
generator performs rule variable substitution only once globally per
rule to put its own placeholders.  Final substitution is performed by
ninja at build time.  Therefore we cannot conditionally replace the
soname placeholders on a per-target basis.

(2) The interface does not define behavior for OS X install-name but it
needs to because CMake implements soname as install-name there.

(3) Some projects may set their own rule variables and depend on
placeholders like

  <CMAKE_SHARED_LIBRARY_SONAME_C_FLAG><TARGET_SONAME>

to work.  We cannot remove the implementation of this substitution.

In order to fix the above we need to refactor the implementation of both
soname and install-name handling.  The flags to set them need to be
moved to the <LINK_FLAGS> placeholder, so the C++ code needs to be
taught exactly when to add the flags.  The placeholders mentioned in
problem (3) must be removed from CMake's platform files, but if a
project adds them they must be honored AND suppress addition of the
flags to <LINK_FLAGS>.