Keep expanding CMAKE_SHARED_LIBRARY_SONAME_${LANG}_FLAG in rules (#13155).
As mentioned in 1a25d0ee point (3), some third party projects might still
depend on rule variables like <CMAKE_SHARED_LIBRARY_SONAME_C_FLAG>. No problem
here since there is absolutely no harm done if we keep expanding it. In case of
<TARGET_SONAME> expansion, it won't happen if NO_SONAME is enabled. However,
that's a new feature which is not enabled by default hence it is rather
reasonable to expect projects to adapt if they want to support it.