Add compiler specific position independent flags.
In almost all cases, this means duplication of
the CMAKE_SHARED_LIBRARY_${lang}_FLAGS for the _PIC
case and using the assumed pie equivalent for the _PIE case.

There is a possibility that the _PIE variables are not correct. However,
as there is no backwards compatibility to be concerned about (as the
POSITION_INDEPENDENT_CODE property is not used anywhere yet), the
current state suffices. The dashboard will find incorrect flags on any
tested platforms.