Port current users of cmGeneratorExpression to two-stage processing.
Removing the Process() API and removing the parameters from the
constructor will allow cmGeneratorExpressions to be cached and
evaluated with multiple configs for example, such as when evaluating
target properties. This requires the creation of a new compiled
representation of cmGeneratorExpression.

The cmListFileBacktrace remains in the constructor so that we can
record where a particular generator expression appeared in the
CMakeLists file.